Create module to know user duplicates
We notice that we could have some user duplications over the People
API.

We want to know if a user could have duplicate users accross the entire
list. To determine if a user is duplicated we are using the gem
'similar_text' which compares two string and give them a score, this
score is on base of 100, where 100 is equal.

We also should considerer implement ourselfs this similar text functionality
consiring that the gem is outdated and looks like nobody is supporting
it anymore.

We create a new class called PeopleDuplicated to manage all the logic for
this requirement.