Link the Duplicates button and the DuplicatePeople service
We need to be able to display the possible duplicated users to
the user that requested it. Currently we have only a button in the
main page, but we need to display the table so the user be aware
of the information.

We are using Ajax to replace the modal of duplicated users. We also
changed the way we were injecting the button into the table, now we
are using a link element with a remote-true attribute which allows us
to call the file duplicated_users.js.erb as part of the response of
the controller.